refactor: extract CalibreLibraryManager from ModelData [P0-1b]
